# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Booklab - backend",
        "description": "Booklab project API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/items": {
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Add new Item",
                "description": "Add Item",
                "operationId": "3c8ff4095ad9d715e12b71a43d40efc2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nome": "Pão"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/items/{id}?with[]=cardapios": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Mostra o item e em qual cardápio ele está presente",
                "description": "Show Item",
                "operationId": "3058b4858d6c10ba68299f25ab92d537",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/items/{id}": {
            "put": {
                "tags": [
                    "Items"
                ],
                "summary": "Updates a Item",
                "operationId": "fc2564bef445e18b2f88e9b0232b5aba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Update Item",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "item_id",
                                "value": "1"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nome": "pão"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Items"
                ],
                "summary": "Delete Items",
                "description": "Delete Items",
                "operationId": "947885a9b94d3ad29b8c5f939da35748",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/reservas": {
            "post": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Add new Reserva",
                "description": "Add Reserva",
                "operationId": "c23f9853525a5a4ad84cd02e1a78944a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "type": "string"
                                    },
                                    "hora": {
                                        "type": "string"
                                    },
                                    "nome_reserva": {
                                        "type": "string"
                                    },
                                    "qtd_pessoas": {
                                        "type": "integer"
                                    },
                                    "contato": {
                                        "type": "string"
                                    },
                                    "cardapio_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "data": "data",
                                    "hora": "hora",
                                    "nome_reserva": "nome",
                                    "qtd_pessoas": 3,
                                    "contato": "contato",
                                    "cardapio_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reservas/{id}?with[]=cardapio": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Mostra a reserva e em qual cardápio será escolhido(opcional)",
                "description": "Show Reserva",
                "operationId": "08f3caffe3110479ba86ec659fc73cb7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/reservas/{id}": {
            "put": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Updates a Reserva",
                "operationId": "32852a4adebca222dda1a1e9938dd697",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Update Reserva",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "item_id",
                                "value": "1"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "type": "string"
                                    },
                                    "hora": {
                                        "type": "string"
                                    },
                                    "nome_reserva": {
                                        "type": "string"
                                    },
                                    "qtd_pessoas": {
                                        "type": "integer"
                                    },
                                    "contato": {
                                        "type": "string"
                                    },
                                    "cardapio_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "data": "data",
                                    "hora": "hora",
                                    "nome_reserva": "nome",
                                    "qtd_pessoas": 3,
                                    "contato": "contato",
                                    "cardapio_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Delete Reserva",
                "description": "Delete Reserva",
                "operationId": "0551814bdb838036eeb0e13e5fcf3b92",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}